+++ server/intvector.h
@@ -2,6 +2,7 @@
 #define __EPOLLSERVER_INTVECTOR_NET_H__
 
 #include <stddef.h>
+#include <stdint.h>
 
 typedef struct {
   int* data;
@@ -12,7 +13,7 @@ typedef struct {
 void intvector_new(IntVector* l, size_t cap);
 void intvector_delete(const IntVector* l);
 void intvector_insert(IntVector* l, int v);
-int intvector_at(const IntVector* l, int i);
+int intvector_at(const IntVector* l, uint64_t i);
 int* intvector_find(const IntVector* l, int v);
 
 #endif
